remove circular dep between abilities and passives
